+++ src/test/kotlin/org/intellij/ml/llm/MyPluginTest.kt
@@ -1,4 +1,4 @@
-package com.github.bzz.intellij
+package org.intellij.ml.llm
 
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.components.service
@@ -6,7 +6,6 @@ import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.util.PsiErrorElementUtil
-import com.github.bzz.intellij.services.MyProjectService
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {
@@ -29,11 +28,11 @@ class MyPluginTest : BasePlatformTestCase() {
         myFixture.testRename("foo.xml", "foo_after.xml", "a2")
     }
 
-    fun testProjectService() {
-        val projectService = project.service<MyProjectService>()
-
-        assertNotSame(projectService.getRandomNumber(), projectService.getRandomNumber())
-    }
+//    fun testProjectService() {
+//        val projectService = project.service<MyProjectService>()
+//
+//        assertNotSame(projectService.getRandomNumber(), projectService.getRandomNumber())
+//    }
 
     override fun getTestDataPath() = "src/test/testData/rename"
 }