+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # intellij Changelog
 
+## 0.0.1 - 2023-10-13
+Internal hackathon version for IJ 2023.2
+
+### Added
+- Initial scaffold from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- server.py, plugin for 2023.2 with ghost/inline code completion provider and a tool window
+
+
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- ...